Make uploading script copy changed files only

+++ upload.sh
@@ -17,6 +17,10 @@ pushd src
             continue
         fi
 
+        if ! git diff --name-only | grep src/$file && [ "$2" != "--force" ] && [ "$file" != "$2" ]; then
+            continue
+        fi
+
         echo "Uploading: src/$file -> $1/$file"
         mkdir -p $(dirname $file)
         cp $file $1/$file
